install.sh: properly handle superuser permissions
This script is based around having superuser permissions. The way it handles them is by using `sudo`. This has two main problems:

1. It could potentially require someone to input their password many, many times depending on their sudo configuration.
2. Even when running as root, the script will break for anyone without sudo installed.

The way to fix this is simple: require the script itself be run as root instead of using sudo. 

This patch adds a check at the beginning which verifies if the user is root. If the user is not root, it explains what must be done to rectify the problem. It then also removes all usages of sudo throughout the script.

This also aligns the script with more typical scriptwriting conventions, overall improving the quality of the script.