install.sh: fail out on command failure
The current setup of install.sh means that all the other commands will still run if any command fails. This creates a volatile and unsafe state should something go wrong, one which may cause recovering difficulties depending on the type of failure.

This patch fixes that problem by adding the "set -e" command to the beginning of the script. This command will cause the script to exit if a command fails.

Signed-off-by: Amy Parker <apark0006@student.cerritos.edu>